refactor: merged all vector sub components into VectorLegend to simplify
