--- sites/geohub/src/components/maplibre/vector/VectorCircle.svelte
@@ -1,75 +1,0 @@
-<script lang="ts">
-	import VectorColorClassification from '$components/maplibre/vector/VectorColorClassification.svelte';
-	import VectorSimulationAccordion from '$components/maplibre/vector/VectorSimulationAccordion.svelte';
-	import type { Tag } from '$lib/types';
-	import {
-		Accordion,
-		CircleRadius,
-		CircleStrokeColor,
-		CircleStrokeWidth,
-		Help,
-		type VectorTileMetadata
-	} from '@undp-data/svelte-undp-components';
-
-	export let layerId: string;
-	export let metadata: VectorTileMetadata;
-	export let tags: Tag[];
-
-	let expanded: { [key: string]: boolean } = {
-		'circle-radius': true
-	};
-	// to allow only an accordion to be expanded
-	let expandedDatasetId: string;
-	$: {
-		let expandedDatasets = Object.keys(expanded).filter(
-			(key) => expanded[key] === true && key !== expandedDatasetId
-		);
-		if (expandedDatasets.length > 0) {
-			expandedDatasetId = expandedDatasets[0];
-			Object.keys(expanded)
-				.filter((key) => key !== expandedDatasetId)
-				.forEach((key) => {
-					expanded[key] = false;
-				});
-			expanded[expandedDatasets[0]] = true;
-		}
-	}
-</script>
-
-<VectorSimulationAccordion {layerId} {tags} bind:expanded />
-
-<Accordion title="Circle radius" bind:isExpanded={expanded['circle-radius']}>
-	<div class="pb-2" slot="content">
-		<CircleRadius {layerId} />
-	</div>
-	<div slot="buttons">
-		<Help>Apply circle radius to the vector layer.</Help>
-	</div>
-</Accordion>
-
-<Accordion title="Circle color" bind:isExpanded={expanded['circle-color']}>
-	<div class="pb-2" slot="content">
-		<VectorColorClassification {layerId} {metadata} propertyName="circle-color" />
-	</div>
-	<div slot="buttons">
-		<Help>Change circle color by using single color or selected property</Help>
-	</div>
-</Accordion>
-
-<Accordion title="Circle stroke color" bind:isExpanded={expanded['circle-stroke-color']}>
-	<div class="pb-2" slot="content">
-		<CircleStrokeColor {layerId} />
-	</div>
-	<div slot="buttons">
-		<Help>The stroke color of the circle.</Help>
-	</div>
-</Accordion>
-
-<Accordion title="Circle stroke width" bind:isExpanded={expanded['circle-stroke-width']}>
-	<div class="pb-2" slot="content">
-		<CircleStrokeWidth {layerId} />
-	</div>
-	<div slot="buttons">
-		<Help>The width of the circle's stroke. Strokes are placed outside of the circle radius.</Help>
-	</div>
-</Accordion>
