--- sites/geohub/src/components/maplibre/vector/VectorPolygon.svelte
@@ -1,53 +1,0 @@
-<script lang="ts">
-	import VectorColorClassification from '$components/maplibre/vector/VectorColorClassification.svelte';
-	import VectorSimulationAccordion from '$components/maplibre/vector/VectorSimulationAccordion.svelte';
-	import type { Tag } from '$lib/types';
-	import {
-		Accordion,
-		FillOutlineColor,
-		Help,
-		type VectorTileMetadata
-	} from '@undp-data/svelte-undp-components';
-
-	export let layerId: string;
-	export let metadata: VectorTileMetadata;
-	export let tags: Tag[];
-
-	let expanded: { [key: string]: boolean } = { 'fill-color': true };
-	// to allow only an accordion to be expanded
-	let expandedDatasetId: string;
-	$: {
-		let expandedDatasets = Object.keys(expanded).filter(
-			(key) => expanded[key] === true && key !== expandedDatasetId
-		);
-		if (expandedDatasets.length > 0) {
-			expandedDatasetId = expandedDatasets[0];
-			Object.keys(expanded)
-				.filter((key) => key !== expandedDatasetId)
-				.forEach((key) => {
-					expanded[key] = false;
-				});
-			expanded[expandedDatasets[0]] = true;
-		}
-	}
-</script>
-
-<VectorSimulationAccordion {layerId} {tags} bind:expanded />
-
-<Accordion title="Fill color" bind:isExpanded={expanded['fill-color']}>
-	<div class="pb-2" slot="content">
-		<VectorColorClassification {layerId} {metadata} propertyName="fill-color" />
-	</div>
-	<div slot="buttons">
-		<Help>Change polygon fill color by using single color or selected property.</Help>
-	</div>
-</Accordion>
-
-<Accordion title="Fill outline color" bind:isExpanded={expanded['fill-outline-color']}>
-	<div class="pb-2" slot="content">
-		<FillOutlineColor {layerId} />
-	</div>
-	<div slot="buttons">
-		<Help>Change polygon outline color.</Help>
-	</div>
-</Accordion>
